+++ src/PESpy/OMF/OMFDBGFile.cs
@@ -37,7 +37,7 @@ namespace PESpy
         /// <inheritdoc/>
         public FileKind Kind => FileKind.OMFDBG;
 
-        public int Length => globalBlock.Length;
+        public long Length => globalBlock.Length;
 
         #region NB02Data
 
@@ -67,7 +67,7 @@ namespace PESpy
             FileName = fileName;
             Name = name ?? Path.GetFileName(fileName);
 
-            globalBlock = new GlobalMemoryBlock(mmf.Address, (int) mmf.Length, this);
+            globalBlock = new GlobalMemoryBlock(mmf.Address, mmf.Length, this);
 
             try
             {
@@ -118,7 +118,7 @@ namespace PESpy
             ILocatorProgress? progress = null,
             CancellationToken cancellationToken = default) => symbolAccessor ??= new NB02SymbolAccessor(data);
 
-        internal unsafe ByteViewProvider CreateByteViewProvider(FileAccessor fileAccessor) => new LocalByteViewProvider(mmf.Address, (int) mmf.Length, fileAccessor);
+        internal unsafe ByteViewProvider CreateByteViewProvider(FileAccessor fileAccessor) => new LocalByteViewProvider(mmf.Address, mmf.Length, fileAccessor);
 
         public void Dispose()
         {
