+++ src/PESpy/OMF/OMFFile.cs
@@ -46,7 +46,7 @@ namespace PESpy
         /// <inheritdoc/>
         public FileKind Kind => FileKind.OMF;
 
-        public int Length => globalBlock.Length;
+        public long Length => globalBlock.Length;
 
         public OMFRecord[] Records { get; }
 
@@ -62,7 +62,7 @@ namespace PESpy
             FileName = fileName;
             Name = name ?? Path.GetFileName(fileName);
 
-            globalBlock = new GlobalMemoryBlock(mmf.Address, (int) mmf.Length, this);
+            globalBlock = new GlobalMemoryBlock(mmf.Address, mmf.Length, this);
 
             try
             {
@@ -130,7 +130,7 @@ namespace PESpy
             ILocatorProgress? progress = null,
             CancellationToken cancellationToken = default) => symbolAccessor ??= new OMFFileSymbolAccessor(this);
 
-        internal unsafe ByteViewProvider CreateByteViewProvider(FileAccessor fileAccessor) => new LocalByteViewProvider(mmf.Address, (int) mmf.Length, fileAccessor);
+        internal unsafe ByteViewProvider CreateByteViewProvider(FileAccessor fileAccessor) => new LocalByteViewProvider(mmf.Address, mmf.Length, fileAccessor);
 
         public void Dispose()
         {
